+++ include/create.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <string>
+
+void createBox(std::string id);
+void waitForInput();
