+++ include/import.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+void importBoxes(bool del);
